pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "folderangel"
version = "1.0.0"
description = "LLM-powered cross-platform folder auto-organizer with an Apple-inspired UI."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "FolderAngel Team"}]
keywords = ["file", "organizer", "llm", "gemini", "pyside6"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PySide6>=6.6",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1",
    "olefile>=0.46",
    "keyring>=24.0",
    "requests>=2.31",
]

[project.optional-dependencies]
windows = ["pywin32>=306; platform_system=='Windows'"]
dev = [
    "pytest>=8",
    "pytest-qt>=4.4",
    "pyinstaller>=6.5",
    "ruff>=0.5",
]

[project.scripts]
folderangel = "folderangel.__main__:main"

[project.gui-scripts]
folderangel-gui = "folderangel.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 110
target-version = "py311"
